+++ webview/src/components/SectionLayout/SectionContent/index.ts
@@ -1,0 +1,5 @@
+import SectionContent from './SectionContent';
+
+export default SectionContent;
+
+export * from './SectionContent';