+++ webview/src/components/SectionLayout/SectionTitle/index.ts
@@ -1,0 +1,5 @@
+import SectionTitle from './SectionTitle';
+
+export default SectionTitle;
+
+export * from './SectionTitle';