+++ webview/src/components/SectionLayout/index.ts
@@ -1,0 +1,5 @@
+import SectionLayout from './SectionLayout';
+
+export default SectionLayout;
+
+export * from './SectionLayout';