+++ webview/src/components/SettingModal/SettingAction/index.ts
@@ -1,0 +1,5 @@
+import SettingAction from './SettingAction';
+
+export default SettingAction;
+
+export * from './SettingAction';