+++ webview/src/components/SettingModal/TranslateSetting/index.ts
@@ -1,0 +1,5 @@
+import TranslateSetting from './TranslateSetting';
+
+export default TranslateSetting;
+
+export * from './TranslateSetting';