+++ webview/src/components/SettingModal/index.ts
@@ -1,0 +1,5 @@
+import SettingModal from './SettingModal';
+
+export default SettingModal;
+
+export * from './SettingModal';